+++ packages/sdk/src/live-chat/live-chat.ts
@@ -29,7 +29,30 @@ export class LiveChat implements BaseLiveChat {
     this.sdk.onUserChanged((user: ExternalUser) => this.watchUserChanged(user));
 
     this.arenaHub = new ArenaHub(chatRoom, sdk);
-    this.arenaHub.track('page');
+
+    if (!this.detectWidgetsPresence()) {
+      this.arenaHub.track('page');
+    }
+  }
+
+  private detectWidgetsPresence() {
+    const arenaLive = document.querySelector('#arena-live');
+    const arenaChat = document.querySelector('#arena-chat');
+    const arenaLiveClass = document.querySelector('.arena-liveblog');
+    const arenaChatClass = document.querySelector('.arena-chat');
+    const arenaWidgetClass = document.querySelector('.arena-embed-widget');
+    const arenaPreviewWidgetClass = document.querySelector('.arena-home-app');
+    const arenaEmbedFrameWidgetClass = document.querySelector('.arena-embed-frame');
+
+    return (
+      !!arenaLive ||
+      !!arenaChat ||
+      !!arenaLiveClass ||
+      !!arenaChatClass ||
+      !!arenaWidgetClass ||
+      !!arenaPreviewWidgetClass ||
+      !!arenaEmbedFrameWidgetClass
+    );
   }
 
   /**